refactor: dependencies graph
